+++ changelog.md
@@ -1,2 +1,2 @@
 ### Wireless Chargers 1.0.9
-- Initial release of Wireless Chargers for Minecraft 1.21.5
+- Initial release of Wireless Chargers for Minecraft 1.21.6 & 1.21.7